+++ src/plugins/vis_type_timeseries/public/application/contexts/vis_data_context.ts
@@ -7,5 +7,6 @@
  */
 
 import React from 'react';
+import { TimeseriesVisData } from 'src/plugins/vis_type_timeseries/common/types';
 
-export const VisDataContext = React.createContext({});
+export const VisDataContext = React.createContext<TimeseriesVisData>({} as TimeseriesVisData);
