+++ src/plugins/vis_type_timeseries/public/application/editor_controller.tsx
@@ -8,37 +8,36 @@
 
 import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
+import { EventEmitter } from 'events';
+
+import { Vis, VisualizeEmbeddableContract } from 'src/plugins/visualizations/public';
+import { IEditorController, EditorRenderProps } from 'src/plugins/visualize/public';
 import { getUISettings, getI18n } from '../services';
 import { VisEditor } from './components/vis_editor_lazy';
+import { TimeseriesVisParams } from '../types';
 
 export const TSVB_EDITOR_NAME = 'tsvbEditor';
 
-export class EditorController {
-  constructor(el, vis, eventEmitter, embeddableHandler) {
-    this.el = el;
+export class EditorController implements IEditorController {
+  constructor(
+    private el: HTMLElement,
+    private vis: Vis<TimeseriesVisParams>,
+    private eventEmitter: EventEmitter,
+    private embeddableHandler: VisualizeEmbeddableContract
+  ) {}
 
-    this.embeddableHandler = embeddableHandler;
-    this.eventEmitter = eventEmitter;
-
-    this.state = {
-      vis: vis,
-    };
-  }
-
-  async render(params) {
+  render({ timeRange, uiState }: EditorRenderProps) {
     const I18nContext = getI18n().Context;
 
     render(
       <I18nContext>
         <VisEditor
           config={getUISettings()}
-          vis={this.state.vis}
-          visParams={this.state.vis.params}
-          timeRange={params.timeRange}
-          renderComplete={() => {}}
-          appState={params.appState}
+          vis={this.vis}
+          timeRange={timeRange}
           embeddableHandler={this.embeddableHandler}
           eventEmitter={this.eventEmitter}
+          uiState={uiState}
         />
       </I18nContext>,
       this.el
