+++ src/plugins/vis_type_timeseries/public/application/lib/fetch_fields.ts
@@ -11,10 +11,12 @@ import { getCoreStart, getDataStart } from '../../services';
 import { ROUTES } from '../../../common/constants';
 import { SanitizedFieldType } from '../../../common/types';
 
+export type VisFields = Record<string, SanitizedFieldType[]>;
+
 export async function fetchFields(
   indexes: string[] = [],
   signal?: AbortSignal
-): Promise<Record<string, SanitizedFieldType[]>> {
+): Promise<VisFields> {
   const patterns = Array.isArray(indexes) ? indexes : [indexes];
   const coreStart = getCoreStart();
   const dataStart = getDataStart();
@@ -32,7 +34,7 @@ export async function fetchFields(
       })
     );
 
-    const fields: Record<string, SanitizedFieldType[]> = patterns.reduce(
+    const fields: VisFields = patterns.reduce(
       (cumulatedFields, currentPattern, index) => ({
         ...cumulatedFields,
         [currentPattern]: indexFields[index],
