+++ src/plugins/vis_type_timeseries/public/types.ts
@@ -8,6 +8,7 @@
 
 import React from 'react';
 import { EuiDraggable } from '@elastic/eui';
+import { PanelSchema } from '../common/types';
 
 type PropsOf<T> = T extends React.ComponentType<infer ComponentProps> ? ComponentProps : never;
 type FirstArgumentOf<Func> = Func extends (arg1: infer FirstArgument, ...rest: any[]) => any
@@ -16,3 +17,5 @@ type FirstArgumentOf<Func> = Func extends (arg1: infer FirstArgument, ...rest: a
 export type DragHandleProps = FirstArgumentOf<
   Exclude<PropsOf<typeof EuiDraggable>['children'], React.ReactElement>
 >['dragHandleProps'];
+
+export type TimeseriesVisParams = PanelSchema;
